Made ItemAPI

+++ ItemAPI/Item.cs
@@ -3,5 +3,5 @@ public class Item
     public long Id { get; set; }
     public string? Title { get; set; }
     public string? Description { get; set;}
-    public decimal unit_price { get; set; } = 0.0m;
+    public double unit_price { get; set; } = 0.0;
 }
